core: Move Particle declaration to separate file
