+++ src/core/serialization/Particle.hpp
@@ -19,10 +19,13 @@
 #ifndef CORE_UTILS_SERIALIZATION_PARTICLE_HPP
 #define CORE_UTILS_SERIALIZATION_PARTICLE_HPP
 
-#include "particle_data.hpp"
+#include "Particle.hpp"
+
 #include <utils/serialization/List.hpp>
 
 #include <boost/serialization/vector.hpp>
+
+class Particle;
 
 namespace boost {
 namespace serialization {
